python-automations/requirements.txt
paho-mqtt
requests
git+https://github.com/csanz91/docker_secrets